Tyuiu.GubanovaSO.Sprint0.Task6.V0.Test/DataServiceTest.cs
using Tyuiu.GubanovaSO.Sprint0.Task6.V0.Lib;
namespace Tyuiu.GubanovaSO.Sprint0.Task6.V0.Test
{
    public class DataServiceTest
    {
        [SetUp]
        public void Setup()
        {
        }

        [Test]
        public void CheckedAdditionArrayValid()
        {
            var numbers = new int [] { 1, 2, 3, 4, 5 };
            var res = DataService.AdditionArray(numbers);
            Assert.AreEqual(15, res);
        }

        [Test]
        public void CheckedSubtractionArrayValid()
        {
            var numbers = new int[] { 1, 2, 3, 4, 5 };
            var res = DataService.SubtractionArray(numbers);
            Assert.AreEqual(-15, res);
        }

        [Test]
        public void CheckeMulArrayValid()
        {
            var numbers = new int[] { 1, 2, 3, 4, 5 };
            var res = DataService.MultiplicationArray(numbers);
            Assert.AreEqual(120, res);
        }
    }
}